requirements.txt
setuptools~=61.3.1
pyinstaller~=5.3
paramiko~=2.11.0
pycurl~=7.45.1
pwntools
impacket~=0.10.0
colorama~=0.4.5
rich~=12.5.0
requests~=2.28.1
mss~=6.1.0
pyfiglet~=0.8.post1
cryptography~=37.0.4
ipapi~=1.0.4
stem~=1.8.0
python-nmap
configparser~=5.3.0
chardet~=5.0.0
ipython~=8.4.0
dnspython~=2.2.1
pysocks

